Syncs player positions once every second.
Figured out dsync problem. If the player is continually pressing the
movement keys the game will send commands across the network every time
the game is updated. If this is the case, the reader gets backed up,
causing a delay that tends to dsync. Even with a one second sync the
player has to not press any keys and allow the reader to read all of the
messages before it can read the one second sync.